L1T vertex finder cleanup : Remove analysis-specific parameters from producer 'cff' file; split 'Settings' class into two, for production and for analysis
