+++ L1Trigger/VertexFinder/interface/L1TrackTruthMatched.h
@@ -19,7 +19,7 @@ class TrackerTopology;
 
 namespace l1tVertexFinder {
 
-class Settings;
+class AnalysisSettings;
 class Stub;
 class TP;
 
@@ -29,7 +29,7 @@ typedef TTClusterAssociationMap<Ref_Phase2TrackerDigi_>        TTClusterAssMap;
 //! Simple wrapper class for TTTrack, with match to a tracking particle
 class L1TrackTruthMatched : public L1Track {
 public:
-  L1TrackTruthMatched(const edm::Ptr<TTTrack< Ref_Phase2TrackerDigi_ >>&, const Settings& , const TrackerGeometry* , const TrackerTopology*, const std::map<edm::Ptr< TrackingParticle >, const TP* >& translateTP, edm::Handle<TTStubAssMap> mcTruthTTStubHandle, edm::Handle<TTClusterAssMap> mcTruthTTClusterHandle, const std::map<DetId, DetId>& geoDetIdMap);
+  L1TrackTruthMatched(const edm::Ptr<TTTrack< Ref_Phase2TrackerDigi_ >>&, const AnalysisSettings& , const TrackerGeometry* , const TrackerTopology*, const std::map<edm::Ptr< TrackingParticle >, const TP* >& translateTP, edm::Handle<TTStubAssMap> mcTruthTTStubHandle, edm::Handle<TTClusterAssMap> mcTruthTTClusterHandle, const std::map<DetId, DetId>& geoDetIdMap);
   ~L1TrackTruthMatched();
   unsigned int getNumStubs()  const  {return numStubs;}
