+++ L1Trigger/VertexFinder/python/VertexProducer_cff.py
@@ -4,49 +4,6 @@ VertexProducer = cms.EDProducer('VertexProducer',
 
   l1TracksInputTag = cms.InputTag("TMTrackProducer", "TML1TracksSimpleLR"), # SFLR
 
-
-
-  #=== Cuts on MC truth particles (i.e., tracking particles) used for tracking efficiency measurements.
-
-  GenCuts = cms.PSet(
-     GenMinPt         = cms.double(3.0),
-     GenMaxAbsEta     = cms.double(2.4),
-     GenMaxVertR      = cms.double(1.0), # Maximum distance of particle production vertex from centre of CMS.
-     GenMaxVertZ      = cms.double(30.0),
-     GenPdgIds        = cms.vuint32(), # Only particles with these PDG codes used for efficiency measurement.
-
-
-     # Additional cut on MC truth tracks used for algorithmic tracking efficiency measurements.
-     # You should usually set this equal to value of L1TrackDef.MinStubLayers below, unless L1TrackDef.MinPtToReduceLayers
-     # is < 10000, in which case, set it equal to (L1TrackDef.MinStubLayers - 1).
-     GenMinStubLayers = cms.uint32(4)
-  ),
-
-
-  #=== Rules for deciding when the track finding has found an L1 track candidate
-
-  L1TrackDef = cms.PSet(
-     UseLayerID           = cms.bool(True),
-     # Reduce this layer ID, so that it takes no more than 8 different values in any eta region (simplifies firmware).
-     ReducedLayerID       = cms.bool(True)
-  ),
-
-  #=== Rules for deciding when a reconstructed L1 track matches a MC truth particle (i.e. tracking particle).
-
-  TrackMatchDef = cms.PSet(
-     #--- Three different ways to define if a tracking particle matches a reco track candidate. (Usually, set two of them to ultra loose).
-     # Min. fraction of matched stubs relative to number of stubs on reco track.
-     MinFracMatchStubsOnReco  = cms.double(-99.),
-     # Min. fraction of matched stubs relative to number of stubs on tracking particle.
-     MinFracMatchStubsOnTP    = cms.double(-99.),
-     # Min. number of matched layers.
-     MinNumMatchLayers        = cms.uint32(4),
-     # Min. number of matched PS layers.
-     MinNumMatchPSLayers      = cms.uint32(0),
-     # Associate stub to TP only if the TP contributed to both its clusters? (If False, then associate even if only one cluster was made by TP).
-     StubMatchStrict          = cms.bool(False)
-  ),
-
   # === Vertex Reconstruction configuration
   VertexReconstruction=cms.PSet(
         # Vertex Reconstruction Algorithm
@@ -78,5 +35,4 @@ VertexProducer = cms.EDProducer('VertexProducer',
     ),
   # Debug printout
   Debug  = cms.uint32(0), #(0=none, 1=print tracks/sec, 2=show filled cells in HT array in each sector of each event, 3=print all HT cells each TP is found in, to look for duplicates, 4=print missed tracking particles by r-z filters, 5 = show debug info about duplicate track removal, 6 = show debug info about fitters)
-  printResults = cms.bool(False)
 )
